In Exec context, open stdin when there's any input
Previously standard in was only being opened when there were argument
or environment variables. If there was a value in the "Input" parameter,
but no arguments or environment variables, an exception was thrown about
stanard input not being redirected. This made it impossible to write
code like

cli {
  Exec "cat"
  Input "foo"
}

To pipe data to a program on stdin.